\name{tagsets}
\alias{Penn_Treebank_POS_tags}
\title{NLP Tag Sets}
\description{
  Tag sets frequently used in Natural Language Processing.
}
\usage{
Penn_Treebank_POS_tags
}
\details{
  \code{Penn_Treebank_POS_tags} provides the Penn Treebank POS tags
  (\url{http://www.cis.upenn.edu/~treebank}) as a data frame with the
  following variables:
  \describe{
    \item{entry}{a character vector with the POS tags}
    \item{description}{a character vector with short descriptions of the
      tags}
    \item{examples}{a character vector with examples for the tags}
  }
}
\examples{
Penn_Treebank_POS_tags
}
